// packages/test/generatedCode/deviceAppManagement/mobileAppConfigurations/item/userStatusSummary/userStatusSummaryRequestBuilderGetRequestConfiguration.ts
import {UserStatusSummaryRequestBuilderGetQueryParameters} from './userStatusSummaryRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface UserStatusSummaryRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: UserStatusSummaryRequestBuilderGetQueryParameters;
}